pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "life-cli"
version = "0.1.0"
description = "Lightweight, stateful, CLI-first orchestrator for personal data pipelines"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Ben Mensi", email = "ben@benthepsychologist.com"}
]
keywords = ["cli", "orchestration", "data-pipeline", "workflow", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "morch>=0.1.0",
    "jinja2>=3.0",
    "llm>=0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
life = "life.cli:main"

[project.urls]
Homepage = "https://github.com/benthepsychologist/life-cli"
Repository = "https://github.com/benthepsychologist/life-cli"
Issues = "https://github.com/benthepsychologist/life-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
life = ["jobs/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
